Lots of small changes. 	modified:   mtrack/database_info.txt 	modified:   mtrack/medicationrecord.cpp 	modified:   mtrack/medicationrecord.h 	modified:   mtrack/medicationsframe.cpp 	modified:   mtrack/medicationsframe.h 	modified:   mtrack/nmw_page00.cpp 	modified:   mtrack/prescriberframe.cpp 	modified:   mtrack/prescriberframe.h 	modified:   mtrack/prescriberframe.ui 	modified:   mtrack/prescriptionsframe.cpp 	modified:   mtrack/prescriptionsframe.ui

+++ mtrack/medicationrecord.cpp
@@ -56,7 +56,7 @@ bool MedicationRecord::retrieve(int newId)
 	manufacturer = model->record(0).value(2).toString();
 	ndc = model->record(0).value(3).toString();
 	form = FORM_INT::strToInt(model->record(0).value(4).toString());
-	strength = model->record(0).value(5).toInt();
+	strength = model->record(0).value(5).toDouble();
 	str_units = model->record(0).value(6).toString();
 	dispense_units = model->record(0).value(7).toString();
 	unit_size = model->record(0).value(8).toString();
